Worked on getting all the classes were on the same page with naming conventions

+++ src/main/java/com/revature/training/BankingApplication/Repository/TransactionRepo.java
@@ -3,5 +3,5 @@ package com.revature.training.BankingApplication.Repository;
 import com.revature.training.BankingApplication.Model.Transactions;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TransactionRepo extends JpaRepository<Transactions, Integer> {
+public interface TransactionRepo extends JpaRepository<Transactions, Long> {
 }
